Threading refactoring.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1562 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: ac2fd952e91003895a18987553ba3f307ff12ef0

+++ Source/MainServer.cpp
@@ -1307,7 +1307,7 @@ static void AdminWork( void* session_ )
                     if( buf.empty() || buf.back() != '\n' )
                         buf += "\n";
 
-                    if( !send_fail && send( s->Sock, buf.c_str(), buf.length() + 1, 0 ) != (int) buf.length() + 1 )
+                    if( !send_fail && send( s->Sock, buf.c_str(), (int) buf.length() + 1, 0 ) != (int) buf.length() + 1 )
                     {
                         WriteLog( ADMIN_PREFIX "Send data fail, disconnect.\n", admin_name );
                         send_fail = true;
